营销服务 第14节:抽奖活动订单流程设计

+++ big-market-infrastructure/src/main/java/com/zhouyihe/bigmarket/infrastructure/persistent/po/RaffleActivityAccount.java
@@ -11,26 +11,48 @@ import java.util.Date;
  */
 @Data
 public class RaffleActivityAccount {
-    /** 主键id */
-    private String id ;
-    /** 用户id */
-    private String userId ;
-    /** 活动id */
-    private Integer activityId ;
-    /** 总次数 */
-    private Integer totalCount ;
-    /** 总次数-剩余 */
-    private Integer totalCountSurplus ;
-    /** 日次数 */
-    private Integer dayCount ;
-    /** 日次数-剩余 */
-    private Integer dayCountSurplus ;
-    /** 月次数 */
-    private Integer monthCount ;
-    /** 月次数-剩余 */
-    private Integer monthCountSurplus ;
-    /** 创建时间 */
-    private Date createTime ;
-    /** 更新时间 */
-    private Date updateTime ;
+    /**
+     * 主键id
+     */
+    private String id;
+    /**
+     * 用户id
+     */
+    private String userId;
+    /**
+     * 活动id
+     */
+    private Long activityId;
+    /**
+     * 总次数
+     */
+    private Integer totalCount;
+    /**
+     * 总次数-剩余
+     */
+    private Integer totalCountSurplus;
+    /**
+     * 日次数
+     */
+    private Integer dayCount;
+    /**
+     * 日次数-剩余
+     */
+    private Integer dayCountSurplus;
+    /**
+     * 月次数
+     */
+    private Integer monthCount;
+    /**
+     * 月次数-剩余
+     */
+    private Integer monthCountSurplus;
+    /**
+     * 创建时间
+     */
+    private Date createTime;
+    /**
+     * 更新时间
+     */
+    private Date updateTime;
 }
